% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_to_aa.R
\name{codons_to_aa}
\alias{codons_to_aa}
\title{Translation of Codons}
\usage{
codons_to_aa(codons)
}
\arguments{
\item{codons}{Vector of codons as strings.}
}
\value{
Returns one string with the amino acid sequence.
}
\description{
Converting codons into amino acids.
}
\examples{
codons <- c("UUU","CAA")
codons_to_aa(codons)

}
