% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_codons.R
\name{divide_codons}
\alias{divide_codons}
\title{Divide into codons}
\usage{
divide_codons(RNA_string, start = 1)
}
\arguments{
\item{start, }{reading frame}
}
\value{
Vector of all the codons included in the string
}
\description{
Divide into codons
}
\examples{

string <- "ATCGTACGATATGATACAGAGATAGACATATTTAACGG"
divide_codons(string)
#"ATC" "GTA" "CGA" "TAT" "GAT" "ACA" "GAG" "ATA" "GAC" "ATA" "TTT" "AAC"
}
