% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dna.R
\name{generate_dna}
\alias{generate_dna}
\title{Generate a random DNA sequence}
\usage{
generate_dna(length)
}
\arguments{
\item{length}{An integer specifying the length of the DNA sequence to
generate.}
}
\value{
A character object with the generated DNA sequence.
}
\description{
Generate a random DNA sequence
}
\examples{
generate_dna(length = 10)
# "CAGGCGCGTT"
}
