% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRandomDNASequence.R
\name{generateRandomDNASequence}
\alias{generateRandomDNASequence}
\title{GenerateRandomDNASequence}
\usage{
generateRandomDNASequence(sequenceLength)
}
\arguments{
\item{sequenceLength}{}
}
\value{
randomDNASequence
}
\description{
GenerateRandomDNASequence
}
