% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aa_occurrence.R
\name{plot_aa_occurrence}
\alias{plot_aa_occurrence}
\title{Plot occurrences of amino acids in peptides}
\usage{
plot_aa_occurrence(protein_string)
}
\arguments{
\item{protein_string}{an amino acid sequence}
}
\value{
Plot of occurrences of each amino acid in a peptide
}
\description{
Plot occurrences of amino acids in peptides
}
\examples{
protein_string <- "GGASVTVSRFW*PSQSKQRHRVEPVS*IQSYLP"
plot_aa_occurrence(protein_string)
}
