% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_occurrence.R
\name{plot_occurrence}
\alias{plot_occurrence}
\title{Plot occurences of amino acids in peptides}
\usage{
plot_occurrence(count_occ)
}
\arguments{
\item{count_occ}{aminoacid string to plot}
}
\value{
plot_of_occ
}
\description{
Plot occurences of amino acids in peptides
}
