% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_plot.R
\name{protein_plot}
\alias{protein_plot}
\title{Plot Amino acids from an input sequence}
\usage{
protein_plot(AA_seq)
}
\arguments{
\item{protein_seq}{Is the input Aminoacid sequence}

\item{unique_protein}{Finds all unique Aminoacids present in the sequence}

\item{protein_plot}{Is the final plot of the Aminoacids are called in the global environment}

\item{visualized_protein}{Is plot of the counts of different Aminoacids.}
}
\value{
the function returns a plot of how many unique protein there are present in the seq.
}
\description{
Plot Amino acids from an input sequence
}
\examples{
protein_plot("AGAAK") #should plot the 3 A's 1 G and 1 K
}
