% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_DNA_sequence.R
\name{Random_DNA_sequence}
\alias{Random_DNA_sequence}
\title{Creating a random DNA sequence}
\usage{
Random_DNA_sequence(sequence_length)
}
\arguments{
\item{sequence_length}{Specifies the length wanted for the random DNA sequence}
}
\value{
A random DNA sequence is created of the length "sequence_length"
}
\description{
Creating a random DNA sequence
}
\examples{
Random_DNA_sequence(10)
}
