% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNA_to_codons.R
\name{RNA_to_codon}
\alias{RNA_to_codon}
\title{Translation of RNA into codons}
\usage{
RNA_to_codon(RNA, start = 1)
}
\arguments{
\item{RNA}{The parameter should be a RNA string, which will be translated to codons}

\item{start}{it tells at what position the translation begins}
}
\value{
The function outputs the condons which have been translated from the RNA
}
\description{
Translation of RNA into codons
}
\examples{
RNA_to_codon("UCCUGG", start=1) #returns c("UCC","UGG")
}
