% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe DNA to RNA, substitute T with U}
\usage{
transcribe(DNA)
}
\arguments{
\item{DNA}{}
}
\value{

}
\description{
Transcribe DNA to RNA, substitute T with U
}
\examples{
transcribe(DNA = "ATTA")
#[1] "AUUA"
}
